fix: Minor adjustments to the example
