+++ aws-greengrass-testing-examples/aws-greengrass-testing-examples-component/pom.xml
@@ -53,6 +53,7 @@
                                       failonerror="true"
                                       newenvironment="true"
                                       classpathref="maven.test.classpath">
+                                    <!-- <sysproperty key="aws.region" value="us-east-1"/> -->
                                     <sysproperty key="ggc.archive" value="greengrass-nucleus-latest.zip"/>
                                     <sysproperty key="gg.component.overrides" value="com.aws.PingPong:file:${project.basedir}/recipe.yaml"/>
                                 </java>
