+++ aws-greengrass-testing-examples/aws-greengrass-testing-examples-component/src/main/java/com/aws/greengrass/testing/examples/component/PaddleComponent.java
@@ -10,5 +10,4 @@ import javax.inject.Singleton;
 public interface PaddleComponent {
     Pong pong();
     Ping ping();
-    EventStreamRPCConnection connection();
 }
