+++ aws-greengrass-testing-examples/aws-greengrass-testing-examples-component/src/main/java/com/aws/greengrass/testing/examples/component/PingPong.java
@@ -1,8 +1,5 @@
 package com.aws.greengrass.testing.examples.component;
 
-import software.amazon.awssdk.eventstreamrpc.EventStreamRPCConnection;
-
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -16,40 +13,15 @@ public class PingPong implements Runnable {
 
     @Override
     public void run() {
-        try (EventStreamRPCConnection connection = component.connection()) {
-            try {
-                CompletableFuture<Void> completed = new CompletableFuture<>();
-                connection.connect(new EventStreamRPCConnection.LifecycleHandler() {
-                    @Override
-                    public void onConnect() {
-                        System.out.println("Connected to IPC.");
-                        completed.complete(null);
-                    }
-
-                    @Override
-                    public void onDisconnect(int i) {
-                    }
-
-                    @Override
-                    public boolean onError(Throwable throwable) {
-                        completed.completeExceptionally(throwable);
-                        return true;
-                    }
-                });
-                completed.get();
-
-                component.ping().watch();
-                component.pong().watch();
-                component.ping().swing();
-
-                // hit back and forth for a few.
-                Thread.sleep(TimeUnit.SECONDS.toMillis(30));
-            } catch (InterruptedException e) {
-                // No-op
-            } catch (TimeoutException | ExecutionException e) {
-                throw new RuntimeException(e);
-            }
-
+        try {
+            component.ping().watch();
+            component.pong().watch();
+            component.ping().swing();
+            Thread.sleep(TimeUnit.SECONDS.toMillis(30));
+        } catch (InterruptedException e) {
+            // No-op
+        } catch (TimeoutException | ExecutionException e) {
+            throw new RuntimeException(e);
         }
     }
