+++ aws-greengrass-testing-examples/aws-greengrass-testing-examples-component/src/test/resources/greengrass/features/ipc.feature
@@ -9,3 +9,6 @@ Feature: Testing a Java Component
       | com.aws.PingPong | LATEST |
     And I deploy the Greengrass deployment configuration
     Then the Greengrass deployment is COMPLETED on the device after 30 seconds
+    And the com.aws.PingPong log on the device contains the line "Connected to IPC" within 20 seconds
+    And the com.aws.PingPong log on the device contains the line "Received Pong from Ping" within 10 seconds
+    And the com.aws.PingPong log on the device contains the line "Received Ping from Pong" within 10 seconds
